requirements.txt
Flask
requests
python-dotenv
google-generativeai
PyPDF2 